requirements.txt
Flask
flask-cors
numpy
pandas
Pillow
opencv-python
scikit-learn
matplotlib
seaborn
werkzeug